add GDS support top mb
