+++ cpp/CMakeLists.txt
@@ -313,6 +313,7 @@ add_library(
   src/io/statistics/parquet_column_statistics.cu
   src/io/text/byte_range_info.cpp
   src/io/text/multibyte_split.cu
+  src/io/text/data_chunk_source_factories.cpp
   src/io/utilities/column_buffer.cpp
   src/io/utilities/config_utils.cpp
   src/io/utilities/data_sink.cpp
