+++ cpp/include/cudf/io/datasource.hpp
@@ -208,7 +208,7 @@ class datasource {
    *
    * @return The number of bytes read (can be smaller than size)
    */
-  virtual size_t device_read(size_t offset, size_t size, uint8_t* dst, rmm::cuda_stream_view stream)
+  virtual size_t device_read(size_t offset, size_t size, void* dst, rmm::cuda_stream_view stream)
   {
     CUDF_FAIL("datasource classes that support device_read must override it.");
   }
@@ -235,7 +235,7 @@ class datasource {
    */
   virtual std::future<size_t> device_read_async(size_t offset,
                                                 size_t size,
-                                                uint8_t* dst,
+                                                void* dst,
                                                 rmm::cuda_stream_view stream)
   {
     CUDF_FAIL("datasource classes that support device_read_async must override it.");
